Let the variables do their jobs
We dont want to hard code the flags here as it will override device BoardConfig.

This will build ROM 4.8 and Kernel 4.7 as default and still allow you to change them if you want. To build with other toolchains add these flags to your BoardConfig

TARGET_GCC_VERSION_AND := (ROM Toolchain to use)
TARGET_GCC_VERSION_ARM := (Kernel toolchain to use)

If a flag is not set, it will go back to the default value